% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablaAnova.R
\name{Tanova}
\alias{Tanova}
\title{Tabla Anova}
\usage{
Tanova(a, b)
}
\arguments{
\item{a}{(vector) variable que representa los datos de la primera raza.}

\item{b}{(vector) variable que representa los datos de la segunda raza.}

\item{data}{base de datos}
}
\value{
un dataframe con la Tabla Anova de un factor.
}
\description{
Realiza una Tabla Anova de un factor para comparar la producción lechera de dos razas de ganado.
}
\examples{
/dontrun{
## Directorio de trabajo
ruta <- "~/ruta/test.csv"

-------------------------------------
# Ejemplo 1:
df <- read.csv(ruta)
# Cargo la libreria
library(AnovaOneF)
Tanova(a = df$HOLSTEIN, b = df$JERSEY)

-------------------------------------
# Ejemplo 2:
a <- c(30, 30, 30, 30, 30, 32, 30, 30, 30, 32, 34, 34, 34, 32, 31)
b <- c(26, 26, 26, 26, 25, 26, 26, 26, 26, 26, 25, 24, 24, 25, 26)

Tanova(a = a, b = b)
}
}
